Moved code from other Util classes to DatabaseUtils
