+++ src/main/resources/application.properties
@@ -13,7 +13,7 @@ spring.flyway.locations=classpath:db/migration/postgres
 #spring.flyway.locations=classpath:db/migration/mysql
 
 ################## Oracle
-# XEPDB1 is service name for plugin database and NOT SID
+## XEPDB1 below is service name for plugin database and NOT SID
 #spring.datasource.jdbc-url=jdbc:oracle:thin:@localhost:1521/XEPDB1
 #spring.datasource.username=jdbctemplatemapper
 #spring.datasource.password=app
